Removal of unnecessary initialization project components
Removal of unnecessary initialization project components
App.spec.tsx - This contains an example UI test that is irrelevant to us right now, and if we do implement UI tests in the future we are likely not going to be using render to test it.
Testing page - this is just an unnecessary page returning just the text "I am a test page" that had a route so I removed that too.